changed type of icon of sidebar item
